+++ dflib/src/main/java/com/nhl/dflib/series/IntSequenceSeries.java
@@ -7,6 +7,7 @@ import com.nhl.dflib.agg.PrimitiveSeriesAvg;
 import com.nhl.dflib.agg.PrimitiveSeriesMedian;
 import com.nhl.dflib.agg.PrimitiveSeriesMinMax;
 import com.nhl.dflib.agg.PrimitiveSeriesSum;
+import com.nhl.dflib.range.Range;
 
 /**
  * An {@link com.nhl.dflib.IntSeries} that represents a range of sequential integers.
@@ -15,23 +16,23 @@ import com.nhl.dflib.agg.PrimitiveSeriesSum;
  */
 public class IntSequenceSeries extends IntBaseSeries {
 
-    private final int first;
-    private final int lastExclusive;
+    private final int fromInclusive;
+    private final int toExclusive;
 
-    public IntSequenceSeries(int first, int lastExclusive) {
+    public IntSequenceSeries(int fromInclusive, int toExclusive) {
 
-        if (first > lastExclusive) {
-            throw new IllegalArgumentException("'first' sequence element is larger than 'last': " + first + " vs. " + lastExclusive);
+        if (fromInclusive > toExclusive) {
+            throw new IllegalArgumentException("'first' sequence element is larger than 'last': " + fromInclusive + " vs. " + toExclusive);
         }
 
-        this.first = first;
-        this.lastExclusive = lastExclusive;
+        this.fromInclusive = fromInclusive;
+        this.toExclusive = toExclusive;
     }
 
     @Override
     public int getInt(int index) {
-        int i = first + index;
-        if (i >= lastExclusive) {
+        int i = fromInclusive + index;
+        if (i >= toExclusive) {
             throw new ArrayIndexOutOfBoundsException(index);
         }
 
@@ -40,12 +41,12 @@ public class IntSequenceSeries extends IntBaseSeries {
 
     @Override
     public void copyToInt(int[] to, int fromOffset, int toOffset, int len) {
-        if (fromOffset + len > lastExclusive) {
+        if (fromOffset + len > toExclusive) {
             throw new ArrayIndexOutOfBoundsException(fromOffset + len);
         }
 
         for (int i = 0; i < len; i++) {
-            to[toOffset + i] = first + i;
+            to[toOffset + i] = fromInclusive + i;
         }
     }
 
@@ -56,11 +57,11 @@ public class IntSequenceSeries extends IntBaseSeries {
             return super.diff(other);
         }
 
-        int otherFirst = ((IntSequenceSeries) other).first;
-        int otherLastExclusive = ((IntSequenceSeries) other).lastExclusive;
+        int otherFirst = ((IntSequenceSeries) other).fromInclusive;
+        int otherLastExclusive = ((IntSequenceSeries) other).toExclusive;
 
-        int head = otherFirst - first;
-        int tail = lastExclusive - otherLastExclusive;
+        int head = otherFirst - fromInclusive;
+        int tail = toExclusive - otherLastExclusive;
 
         if (head > 0) {
             if (tail <= 0) {
@@ -70,10 +71,10 @@ public class IntSequenceSeries extends IntBaseSeries {
             int len = head + tail;
             int[] data = new int[len];
             for (int i = 0; i < head; i++) {
-                data[i] = first + i;
+                data[i] = fromInclusive + i;
             }
 
-            int tailOffset = lastExclusive - tail - 1;
+            int tailOffset = toExclusive - tail - 1;
             for (int i = head; i < len; i++) {
                 data[i] = tailOffset + i;
             }
@@ -93,11 +94,11 @@ public class IntSequenceSeries extends IntBaseSeries {
             return super.intersect(other);
         }
 
-        int otherFirst = ((IntSequenceSeries) other).first;
-        int otherLastExclusive = ((IntSequenceSeries) other).lastExclusive;
+        int otherFirst = ((IntSequenceSeries) other).fromInclusive;
+        int otherLastExclusive = ((IntSequenceSeries) other).toExclusive;
 
-        int head = otherFirst - first;
-        int tail = lastExclusive - otherLastExclusive;
+        int head = otherFirst - fromInclusive;
+        int tail = toExclusive - otherLastExclusive;
 
         if (head > 0) {
             return tail > 0 ? head(-head).tail(-tail) : head(-head);
@@ -115,9 +116,17 @@ public class IntSequenceSeries extends IntBaseSeries {
 
     @Override
     public IntSeries rangeOpenClosedInt(int fromInclusive, int toExclusive) {
-        return fromInclusive == 0 && toExclusive == size()
-                ? this
-                : new IntSequenceSeries(first + fromInclusive, lastExclusive + toExclusive);
+        int len = size();
+        if (fromInclusive == 0 && toExclusive == len) {
+            return this;
+        }
+
+        int newLen = toExclusive - fromInclusive;
+        int newFrom = this.fromInclusive + fromInclusive;
+        int newTo = this.fromInclusive + toExclusive;
+
+        Range.checkRange(fromInclusive, newLen, len);
+        return new IntSequenceSeries(newFrom, newTo);
     }
 
     @Override
@@ -127,52 +136,53 @@ public class IntSequenceSeries extends IntBaseSeries {
             return this;
         }
 
-        return len < 0 ? tail(size() + len) : new IntSequenceSeries(first, first + len);
+        return len < 0 ? tail(size() + len) : rangeOpenClosedInt(0, len);
     }
 
     @Override
     public IntSeries tail(int len) {
+        int size = size();
 
         if (Math.abs(len) >= size()) {
             return this;
         }
 
-        return len < 0 ? head(size() + len) : new IntSequenceSeries(first + size() - len, lastExclusive);
+        return len < 0 ? head(size + len) : rangeOpenClosedInt(size - len, size);
     }
 
     @Override
     public int size() {
-        return lastExclusive - first;
+        return toExclusive - fromInclusive;
     }
 
     @Override
     public int max() {
-        return PrimitiveSeriesMinMax.maxOfRange(lastExclusive);
+        return PrimitiveSeriesMinMax.maxOfRange(toExclusive);
     }
 
     @Override
     public int min() {
-        return PrimitiveSeriesMinMax.minOfRange(first);
+        return PrimitiveSeriesMinMax.minOfRange(fromInclusive);
     }
 
     @Override
     public long sum() {
-        return PrimitiveSeriesSum.sumOfRange(first, lastExclusive);
+        return PrimitiveSeriesSum.sumOfRange(fromInclusive, toExclusive);
     }
 
     @Override
     public double avg() {
-        return PrimitiveSeriesAvg.avgOfRange(first, lastExclusive);
+        return PrimitiveSeriesAvg.avgOfRange(fromInclusive, toExclusive);
     }
 
     @Override
     public double median() {
-        return PrimitiveSeriesMedian.medianOfRange(first, lastExclusive);
+        return PrimitiveSeriesMedian.medianOfRange(fromInclusive, toExclusive);
     }
 
     @Override
     public LongSeries cumSum() {
-        long[] cumSum = PrimitiveSeriesSum.cumSumOfRange(first, lastExclusive);
+        long[] cumSum = PrimitiveSeriesSum.cumSumOfRange(fromInclusive, toExclusive);
         return new LongArraySeries(cumSum);
     }
 }